requirements/base.txt
python-dateutil==2.6.0
pytz==2018.3
tzlocal==1.5.1

Django==2.0.8
psycopg2==2.7
django-braces==1.13.0
django-adminactions==1.5
django-suit==0.2.26
django-bootstrap3==11.0.0
django-select2==6.1.0
djangorestframework==3.8.2
django-sql-explorer==1.1.1
django-paintstore==0.2
django-solo==1.1.3
django-extensions==2.1.4
django-filter==2.0.0
django-rest-swagger==2.1.2
django-crispy-forms==1.7.2
djangorestframework-bulk==0.2.1
django-countries==4.5
django-localflavor==1.3
django-webpack-loader==0.6.0
django-mass-edit==3.1
django-fobi==0.13.2

# Django admin docs
docutils==0.12
requests==2.19.1

# PDF
xhtml2pdf==0.2.2
pdfminer.six==20181108

# PIL (Image Upload)
Pillow==5.2.0

# static wsgi extensions
dj-static==0.0.6
static3==0.5.1

# rest api
drf-nested-routers==0.11.1
# web access request
netaddr==0.7.18

# Cron jobs
plan==0.5

django-easy-pdf==0.2.0-dev1

ortools==6.5.4527
